avm2: Significantly improve optimizer
