+++ core/src/avm2/optimizer/mod.rs
@@ -1,0 +1,4 @@
+mod blocks;
+mod optimize;
+
+pub use optimize::optimize;
